perf: Massively reduce memory usage
Previously, the entire mosaic was built in memory before encoding to
PNG, resulting in up to 1.3 GB of memory usage since the full image is
very large (16384x12288 in most cases).

With this change, rows of tiles are downloaded as needed during encoding.
`image.Image` doesn't typically work this way in Go, so there's a chance
this could break in a future Go version.